Add home page and lasagna recipe
Created an index.html page to be able to reference
multiple recipe pages. Created a page for a
lasagna recipe and added steps and ingredients.
Linked the pages together so that it's possible
to navigate between them. Updated README.md page
to include some information on what this project
is for.